Change the XML configuration paths to be converted to absolute paths based on the directory of the XML file instead of the current working directory.
It makes the configuration paths more predictable.

Before the patch, you were forced to run the phpunit command from the same directory as the XML configuration file.
For instance, to run the PHPUnit tests, you must use the following:

  cd PHPUnit/Tests/
  phpunit --configuration=configuration.xml

After the patch, the above still works, but you can also run the tests from the main PHPUnit directory:

  cd ../..
  phpunit --configuration=PHPUnit/Tests/configuration.xml

The patch also changes the paths contained in the current test suite of PHPUnit, as it assumes that PHPUnit is run from the XML file directory.