Added the ability to configure the autoloader via using the configuration file.
Individual namespace and prefix configurations are given a name and a set of
paths to directories, each of which may contain the source file that needs to be
loaded.

The `PHPUnit_Util_Configuration#getAutoloaderConfiguration` method returns an
associative array, which contains both the namespace and prefix configurations
under the "namespace" and "prefix" keys respectively.

The `PHPUnit_TextUI_Command#main` method then hooks all this together:
configuring and registering a new instance of the `PHPUnit_Util_Autoloader`
class when it is required.