Quick hack to make ResultPrinter.php include excluded tests in its output.
I choose to use an underscore as the output as we would want to have the total percentage complete to end up with 100%. If instead we would call writeProgress(NULL) and not have it output anything nor increment $this->column, it would cause the total summary to continue to increment. That does have the side effect where if you hit a chunk of excluded tests, your counter may go from 122/1300 to 553/1300 on the following line with no real indication of why.

Another option I thought of was to instead cause the total number of tests counter to decrement, but that also looks weird if you go from 120/1500 (8%) to 160/800 (20%). But in the case when you exclude large chunks, it would prevent lots of useless output.