files in testsuite are now loaded using an absolute path
When using a wrapper around PHPUnit, the files defined in testsuites
might not be found when declared as a relative path.  This is was caused
by PHPUnit not normalizating them to absolute paths although directories
are normalized.

Use case is the MediaWiki test suite which has tests materials under
/tests/phpunit.  There is a phpunit.php wrapper and the suite.xml file
definining a testsuite with <file>suites/ExtensionsTestsuite.php</file>.
When invoking the wrapper in the samedir as the suite.xml, the relative
path is obviously found.  When invoking the wrapper from a parent
directory, it would fail to find it.