Ability to call ::method() on a stub without calling ::expects()
When using a Test Double as a Stub, we don't typically care whether it is called. Currently the mocking library forces the test author to specify an expectation of how many times the method will be called. This makes it harder to visually scan source code to see which Doubles have expectations set, and which don't care.

This change allows calls like $this->getMock()->method('foo')->will($this->returnValue('bar')); which are clearer to identify as Stubbing behaviour.