Support filters when only generating XML coverage
Currently filters are ignored if XML coverage is the only coverage being used.

NOTE: This patch also fixes an apparent logic bug; please verify this is a bug before merging. The current code only checks `this->canCollectCodeCoverage` if $arguments['coverageText'] is true, which seemed like a bug to me. If this is intentional, I suggest renaming `canCollectCodeCoverage` to `canCollectTextCodeCoverage`.