Add another possible autoload location
When using composer, and not symlinked, this binary ended up in `[project]/bin` with the autoloader at `vendor/autoload.php`, so the path needed is just `../vendor/autoload.php`. I'm not sure where this script would ever end up that required `../../` as the path - it may be safe to delete line 41.