Fix JsonMatches for [] and {}
This commit fixes #2919 (assertJsonStringEqualsJsonString empty object
matches empty array).

We had a problem where JsonMatches would report that the JSON string
'{}' matches the JSON string '[]'. This bug was due to the fact that
Json::canonicalize() was converting JSON strings to PHP arrays. In a PHP
array, there is no difference between an empty associative array and an
empty indexed array, so Json::canonicalize() did not distinguish between
them either. The solution we apply here is to deserialize JSON into
objects (rather than associative arrays) so that PHP _can_ distinguish
between an empty object and an empty array. If the object is not empty,
we still ultimately convert it to an associative array so we can sort
its keys.