Fix StringMatchesFormatDescription with \r\n
This is a fix for #3040.

There was a bug in that the following assertion would fail, even though
it should pass:

    $this->assertStringMatchesFormat("\r\n", "\r\n");

Obviously, we would expect a string to match itself. The bug was due to
newline conversion - `\r\n` was being converted to `\n` on the pattern,
but not on the string to be matched against.

In order to be compatible with existing behavior, we need to continue
converting `\r\n` to `\n` in the pattern. (Otherwise, users who expect
this behavior may have tests fail that would have previously passed.)
Therefore, the only option for fixing the bug is to also sanitize the
string to be matched. This commit makes that change, and provides some
additional tests to verify the behavior.