Throw exception when ignoreCase on array
The documentation for assertContains() has 2 forms:

    assertContains(mixed $needle, Iterator|array $haystack[, string $message = ''])

and

    assertContains(string $needle, string $haystack[, string $message = '', boolean $ignoreCase = false])

Some users may become confused and try to pass `$ignoreCase=true` with
the first form. In this case, PHPUnit was silently ignoring the
parameter. Rather than silently ifnoring the parameter, we should throw
an exception to alert the user that this is an invalid parameter
combination.