Use finally insead of isset on error
By using the `finally` keyword we don't need to check if the error
was thrown, and then throw it again after the checks.

This _should_ also help a bit more in case one needs to debug where
an error came from, as the original error is still raised, instead of
being caught and thrown again.