CI: make phars created during build available
I wanted to try and test something with the latest `dev-9.5` version (against PHP 8.1) and realized I would have to build the phar myself.

To make it easier for people to test proposed changes or the latest dev version of PHPUnit with a phar, I'd like to propose to automatically make the phar(s) generated during the CI builds available as artifacts.

Per the discussion in the PR:
* For each build one (1) phar file will be uploaded, which will be available as an artifact in a zip file.
* This is the `scoped` phar.
* And will be retained for 7 days.

The zips can be downloaded from the Actions Summary page after each CI run.

Also see: https://github.com/actions/upload-artifact