Better fix for #4740
Instead of replacing static loading with autoloading we now combine the best of both worlds. The PHAR now loads all sourcecode files bundled in it on startup (again), but assisted by an autoloader so that PHP's compiler can find code units that have not been loaded yet for perming method compatibility checks, for instance.