Use spl_object_id instead of spl_object_hash where possible.
spl_object_id was added in php 7.2.
array_push allows the empty array since php 7.3.
Modify array in place rather than copying all values to new arrays
repeatedly.

Integers are slightly faster to compare than strings and use less
memory.

This in particular isn't performance sensitive, but update this as an
example for future additions or code based on it.