Constraint/logicalNot: bug fix - prevent weird replacement
If a value passed to an assertion would have `exists` in it, it would get replaced by `does not exist` leading to weird error messages.

Fixed now.

I have a test for this, but as this code isn't tested in the 8.5 branch, I will pull the test separately and to the `main` branch.