GH Actions: actually run the tests on Windows
PR 3982 added test run builds against Windows OS.

I'm not sure if this ever worked, but it sure isn't working now. The environment is created, but the actual steps to run things are not doing anything due to the command not being understood by Windows.

This commit fixes that and allows the tests to actually run on Windows.