Change Constraint::toString() to conditionally export objects
Objects are exported by default when toString() is called from
failureDescription(), for example, to create assertion failure messages.

Objects are not exported by default when toString() is called outside
the context of assertion failure message creation. Most notably, objects
are not exported by default when toString() is called from the event
emitter.