[TASK] Provide a test to demonstrate a regresision
This change adds a fixture class and a MockBuilder test
to showcase a regression introduced with version 11.2.0.

Given is a class, where one method is mocked using the
`onlyMethods()` option to check the method gets called.
The method is called in the class constructor and also
in other public method.

Providing a test mocking that method this breaks since
11.2.0 - albeit having the count in previous version
only working for calls to the mock method AFTER the
constructor call.

Following exception thrown is not expectd:

    Error: Typed property
    MockObject_ClassCallingMethodInConstructor_b5ada611::$__phpunit_state
    must not be accessed before initialization

Related: #5857