fix preg_match() warning for most filter usecases
when a --filter option is provided and it starts with an invalid
regex-delimiter, it will emit the following warning:

  Warning: preg_match(): Delimiter must not be alphanumeric, backslash, or NUL

this is really cumbersome when xdebug.scream=true is in use. also while
stepdebugging with phpstorm for example, this will spam the console
output of phpstorm.

we can and should prevent that most obvious case by just checking if the
supplied filter starts with an alphanumeric character.